KOGITO-7373 WIP - moved to knative-serving addon
