+++ quarkus/addons/knative/serving/deployment/src/main/resources/META-INF/services/org.kie.kogito.serverless.workflow.parser.FunctionTypeHandler
@@ -1,0 +1,1 @@
+org.kie.kogito.addons.quarkus.knative.serving.deployment.customfunctions.KnativeTypeHandler
