+++ quarkus/addons/knative/serving/pom.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>org.kie.kogito</groupId>
+        <artifactId>kogito-addons-quarkus-knative-parent</artifactId>
+        <version>2.0.0-SNAPSHOT</version>
+    </parent>
+
+    <artifactId>kogito-addons-quarkus-knative-serving-parent</artifactId>
+    <name>Kogito Add-On Knative Serving - Parent</name>
+    <modules>
+        <module>deployment</module>
+        <module>runtime</module>
+    </modules>
+    <packaging>pom</packaging>
+
+</project>
