+++ quarkus/addons/knative/serving/runtime/src/main/resources/META-INF/kogito.addon
@@ -1,0 +1,1 @@
+knative-serving
