+++ quarkus/addons/knative/serving/runtime/src/test/java/org/kie/kogito/addons/quarkus/knative/serving/customfunctions/KnativeServiceDiscoveryTest.java
@@ -1,0 +1,86 @@
+/*
+ * Copyright 2022 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.kie.kogito.addons.quarkus.knative.serving.customfunctions;
+
+import java.util.Optional;
+
+import javax.inject.Inject;
+
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import io.fabric8.knative.client.KnativeClient;
+import io.fabric8.knative.serving.v1.Service;
+import io.fabric8.kubernetes.client.server.mock.KubernetesServer;
+import io.quarkus.test.junit.QuarkusTest;
+import io.quarkus.test.kubernetes.client.KubernetesTestServer;
+import io.quarkus.test.kubernetes.client.WithKubernetesTestServer;
+
+@QuarkusTest
+@WithKubernetesTestServer
+class KnativeServiceDiscoveryTest {
+
+    @KubernetesTestServer
+    KubernetesServer mockServer;
+
+    @Inject
+    KnativeServiceDiscovery knativeServiceDiscovery;
+
+    static KnativeClient knativeClient;
+
+    @BeforeEach
+    void createServiceIfNotExists() {
+        String namespace = "test";
+
+        if (mockServer.getClient().services().inNamespace("test").withName("serverless-workflow-greeting-quarkus").get() != null) {
+            return;
+        }
+
+        knativeClient = mockServer.getClient().adapt(KnativeClient.class);
+        Service kService = knativeClient.services()
+                .inNamespace(namespace)
+                .load(getClass().getClassLoader().getResourceAsStream("knative/quarkus-greeting.yaml"))
+                .get();
+        knativeClient.services().inNamespace(namespace).resource(kService).create();
+    }
+
+    @AfterAll
+    static void cleanup() {
+        knativeClient.close();
+    }
+
+    @Test
+    void discoverSpecificNamespace() {
+        Optional<KnativeServiceServer> server = knativeServiceDiscovery.discover("test/serverless-workflow-greeting-quarkus");
+        Assertions.assertThat(server).map(KnativeServiceServer::getHost).hasValue("serverless-workflow-greeting-quarkus.test.10.99.154.147.sslip.io");
+        Assertions.assertThat(server).map(KnativeServiceServer::getPort).hasValue(80);
+    }
+
+    @Test
+    void discoverCurrentNamespace() {
+        Optional<KnativeServiceServer> server = knativeServiceDiscovery.discover("serverless-workflow-greeting-quarkus");
+        Assertions.assertThat(server).map(KnativeServiceServer::getHost).hasValue("serverless-workflow-greeting-quarkus.test.10.99.154.147.sslip.io");
+        Assertions.assertThat(server).map(KnativeServiceServer::getPort).hasValue(80);
+    }
+
+    @Test
+    void serviceInDifferentNamespaceShouldNotBeFound() {
+        Optional<KnativeServiceServer> server = knativeServiceDiscovery.discover("different_namespace/serverless-workflow-greeting-quarkus");
+        Assertions.assertThat(server).isEmpty();
+    }
+}
