+++ quarkus/addons/knative/serving/runtime/src/test/java/org/kie/kogito/addons/quarkus/knative/serving/customfunctions/KnativeServiceTest.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2022 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.kie.kogito.addons.quarkus.knative.serving.customfunctions;
+
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+class KnativeServiceTest {
+
+    @Test
+    void testOnlyService() {
+        KnativeService knativeService = new KnativeService("service");
+
+        assertThat(knativeService.getNamespace()).isEmpty();
+        assertThat(knativeService.getName()).isEqualTo("service");
+    }
+
+    @Test
+    void testNamespaceAndService() {
+        KnativeService knativeService = new KnativeService("namespace/service");
+
+        assertThat(knativeService.getNamespace()).hasValue("namespace");
+        assertThat(knativeService.getName()).isEqualTo("service");
+    }
+}
