--- quarkus/addons/kubernetes/deployment/src/main/java/org/kie/kogito/addons/quarkus/kubernetes/functions/knative/KnativeTypeHandler.java
@@ -1,43 +1,0 @@
-/*
- * Copyright 2022 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.kie.kogito.addons.quarkus.kubernetes.functions.knative;
-
-import org.jbpm.ruleflow.core.RuleFlowNodeContainerFactory;
-import org.jbpm.ruleflow.core.factory.WorkItemNodeFactory;
-import org.kie.kogito.addons.quarkus.k8s.functions.knative.KnativeWorkItemHandler;
-import org.kie.kogito.serverless.workflow.parser.ParserContext;
-import org.kie.kogito.serverless.workflow.parser.types.WorkItemTypeHandler;
-
-import io.serverlessworkflow.api.Workflow;
-import io.serverlessworkflow.api.functions.FunctionDefinition;
-
-import static org.kie.kogito.serverless.workflow.parser.FunctionTypeHandlerFactory.trimCustomOperation;
-
-public class KnativeTypeHandler extends WorkItemTypeHandler {
-
-    @Override
-    protected <T extends RuleFlowNodeContainerFactory<T, ?>> WorkItemNodeFactory<T> fillWorkItemHandler(Workflow workflow,
-            ParserContext context,
-            WorkItemNodeFactory<T> node,
-            FunctionDefinition functionDef) {
-        return node.workName(KnativeWorkItemHandler.NAME).metaData("operation", trimCustomOperation(functionDef));
-    }
-
-    @Override
-    public String type() {
-        return "knative";
-    }
-}
