--- quarkus/addons/kubernetes/runtime/src/main/java/org/kie/kogito/addons/quarkus/k8s/functions/knative/KnativeWorkItemHandlerConfig.java
@@ -1,34 +1,0 @@
-/*
- * Copyright 2022 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.kie.kogito.addons.quarkus.k8s.functions.knative;
-
-import javax.annotation.PostConstruct;
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
-
-import org.kie.kogito.process.impl.DefaultWorkItemHandlerConfig;
-
-@ApplicationScoped
-final class KnativeWorkItemHandlerConfig extends DefaultWorkItemHandlerConfig {
-
-    @Inject
-    KnativeWorkItemHandler handler;
-
-    @PostConstruct
-    void init() {
-        register(handler.getName(), handler);
-    }
-}
