--- quarkus/addons/kubernetes/runtime/src/main/resources/META-INF/services/org.kie.kogito.serverless.workflow.parser.FunctionTypeHandler
@@ -1,1 +1,0 @@
-org.kie.kogito.addons.quarkus.kubernetes.functions.knative.KnativeTypeHandler
